requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
requests==2.31.0
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.10.1
isort==5.12.0
flake8==6.1.0 